# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TrendPoint {
  label: string;
  risk: number;
  volume: number;
}

export interface RiskOverview {
  transactionsAnalyzed: number;
  highRisk: number;
  fraudDetected: number;
  preventedLoss: number;
  activeInvestigations: number;
  averageRiskScore: number;
  fraudRate: number;
  spikeStatus: string;
  trend: TrendPoint[];
}

export interface RiskTransaction {
  transactionId: string;
  customerId: string;
  merchantId: string;
  amount: number;
  currency: string;
  timestamp: string;
  riskScore: number;
  riskLevel: string;
  fraudProbability: number;
  anomalyScore: number;
  behaviorScore: number;
  velocityScore: number;
  graphScore: number;
  decision: string;
  status: string;
  factors: string[];
}

export interface EvidenceItem {
  label: string;
  value: string;
  source: string;
}

export interface Investigation {
  transaction: RiskTransaction;
  summary: string;
  confidence: number;
  recommendation: string;
  missingInformation: string[];
  facts: string[];
  inferences: string[];
  recommendations: string[];
  evidence: EvidenceItem[];
}

export interface NetworkNode {
  id: string;
  label: string;
  type: string;
  risk: number;
}

export interface NetworkLink {
  source: string;
  target: string;
}

export interface RiskNetwork {
  clusterCount: number;
  nodes: NetworkNode[];
  links: NetworkLink[];
}

export type ReviewDecisionInputDecision = typeof ReviewDecisionInputDecision[keyof typeof ReviewDecisionInputDecision];


export const ReviewDecisionInputDecision = {
  approve: 'approve',
  reject: 'reject',
  escalate: 'escalate',
  request_evidence: 'request_evidence',
} as const;

export interface ReviewDecisionInput {
  decision: ReviewDecisionInputDecision;
  note?: string;
}

export interface AuditEvent {
  id: string;
  caseId: string;
  timestamp: string;
  event: string;
  actor: string;
  note?: string;
  decisionVersion: string;
}