# lib/api-zod/src/generated/types/investigation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceItem } from './evidenceItem';
import type { RiskTransaction } from './riskTransaction';

export interface Investigation {
  transaction: RiskTransaction;
  summary: string;
  confidence: number;
  recommendation: string;
  missingInformation: string[];
  facts: string[];
  inferences: string[];
  recommendations: string[];
  evidence: EvidenceItem[];
}