# lib/api-zod/src/generated/types/networkNode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NetworkNode {
  id: string;
  label: string;
  type: string;
  risk: number;
}