# lib/api-zod/src/generated/types/reviewDecisionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewDecisionInputDecision } from './reviewDecisionInputDecision';

export interface ReviewDecisionInput {
  decision: ReviewDecisionInputDecision;
  note?: string;
}