# lib/api-zod/src/generated/types/riskNetwork.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NetworkLink } from './networkLink';
import type { NetworkNode } from './networkNode';

export interface RiskNetwork {
  clusterCount: number;
  nodes: NetworkNode[];
  links: NetworkLink[];
}