# lib/api-zod/src/generated/types/riskOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrendPoint } from './trendPoint';

export interface RiskOverview {
  transactionsAnalyzed: number;
  highRisk: number;
  fraudDetected: number;
  preventedLoss: number;
  activeInvestigations: number;
  averageRiskScore: number;
  fraudRate: number;
  spikeStatus: string;
  trend: TrendPoint[];
}