# lib/api-zod/src/generated/types/riskTransaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RiskTransaction {
  transactionId: string;
  customerId: string;
  merchantId: string;
  amount: number;
  currency: string;
  timestamp: string;
  riskScore: number;
  riskLevel: string;
  fraudProbability: number;
  anomalyScore: number;
  behaviorScore: number;
  velocityScore: number;
  graphScore: number;
  decision: string;
  status: string;
  factors: string[];
}